+++ changelog.md
@@ -24,6 +24,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0)
   If this slowdown in e.g. `tap-macro`s bothers you, you may want to set it to 0.
   If you used `cmp-seq-delay`, you probably no longer have to,
   as `key-seq-delay` effectivly already implies a `cmp-seq-delay`.
+- Any delay may now be fractional up to a resolution of microseconds.
+  That is up to three extra fractional digits. (#975)
 
 ### Fixed